refactor accordion
