+++ examples/angular/src/app/accordionWrapper/accordion.component.ts
@@ -1,12 +1,10 @@
-import { Component, CUSTOM_ELEMENTS_SCHEMA, ElementRef, input, signal } from '@angular/core';
-import { AccordionItemComponent } from '../accordionItemWrapper/accordionItem.component';
+import { Component, CUSTOM_ELEMENTS_SCHEMA, ElementRef, input } from '@angular/core';
 
 @Component({
   selector: 'app-accordion',
   standalone: true,
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   templateUrl: './accordion.component.html',
-  imports: [AccordionItemComponent],
 })
 export class AccordionComponent {
   constructor(elementRef: ElementRef) {
@@ -17,19 +15,4 @@ export class AccordionComponent {
   bordered = input.required<string>();
   items = input<{ title: string; content: string }[]>();
   headingLevel = input<string>();
-
-  liked = signal(true);
-
-  setLiked(e: Event) {
-    const value = ((e as any).detail.target as HTMLInputElement).value === 'true';
-    this.liked.set(value);
-  }
-
-  alertLiked() {
-    if (this.liked()) {
-      alert('Glad you liked this!');
-    } else {
-      alert('Sorry you did not like this content.');
-    }
-  }
 }
