+++ examples/angular/src/app/app.component.html
@@ -18,7 +18,64 @@
 
     <!-- Wrapped Accordion -->
     <h2 class="ds-text-heading--4xl ds-u-margin-top--2">Accordion</h2>
-    <app-accordion [bordered]="bordered" [headingLevel]="headingLevel" />
+    <app-accordion [bordered]="bordered" [headingLevel]="headingLevel">
+      <app-accordion-item>
+        <item-heading>The First Ammendment</item-heading>
+        <ds-alert> This is an alert inside of an accordion item. </ds-alert>
+        <p>
+          The First Amendment provides several rights protections: to express ideas through speech
+          and the press, to assemble or gather with a group to protest or for other reasons, and to
+          ask the government to fix problems. It also protects the right to religious beliefs and
+          practices. It prevents the government from creating or favoring a religion.
+        </p>
+        <div class="ds-u-margin-top--2">
+          <ds-choice-list
+            (ds-change)="setLiked($event)"
+            class-name="ds-u-margin-bottom--2"
+            label="Did you like this content?"
+            name="likedIt"
+            type="radio"
+          >
+            <ds-choice label="Yes" default-checked="true" value="true" />
+            <ds-choice label="No" value="false" />
+          </ds-choice-list>
+          <ds-button (ds-click)="alertLiked()" variation="solid"
+            >Tell us what you though about this content.</ds-button
+          >
+        </div>
+      </app-accordion-item>
+      <app-accordion-item>
+        <item-heading>The Third Ammendment</item-heading>
+        <ds-alert variation="warn">
+          This is an alert inside of an accordion item that is warning you.</ds-alert
+        >
+        <p>
+          The Third Amendment prevents government from forcing homeowners to allow soldiers to use
+          their homes. Before the Revolutionary War, laws gave British soldiers the right to take
+          over private homes.
+        </p>
+      </app-accordion-item>
+      <app-accordion-item>
+        <item-heading>The Fourth Ammendment</item-heading>
+        <ds-alert variation="error">
+          This is an alert inside of an accordion item that informing you of an error.</ds-alert
+        >
+        <p>
+          The Fourth Amendment bars the government from unreasonable search and seizure of an
+          individual or their private property.
+        </p>
+      </app-accordion-item>
+      <div class="ds-u-margin-top--2">
+        <cite>
+          <a
+            href="https://www.archives.gov/founding-docs/bill-of-rights/what-does-it-say"
+            target="_blank"
+          >
+            Source Material</a
+          >
+        </cite>
+      </div>
+    </app-accordion>
 
     <h2 class="ds-text-heading--4xl ds-u-margin-top--2">Month Picker</h2>
     <ds-month-picker
